pyproject.toml
[tool.poetry]
name = "toku-superhero-simulation"
version = "0.1.0"
description = ""
authors = ["Saúl Hormazábal <saul.hormazabal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.7"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
setuptools = "^69.0.2"
celery = "^5.3.6"
flower = "^2.0.1"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
django-pgviews-redux = "^0.9.4"
django-filter = "^23.4"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pdbpp = "^0.10.3"
django-debug-toolbar = "^4.2.0"
django-extensions = "^3.2.3"
prospector = "^1.10.3"
pylint = "2.13.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
